Add basic selenium tests
Can't test yet, but should be similar to before, just running " pytest " in the terminal to run them. Only have login and post right now.
