+++ src/requirements.txt
@@ -12,3 +12,4 @@ typing_extensions==4.10.0
 Werkzeug==3.0.1
 flask-wtf==1.2.1
 flask_login==0.6.3
+selenium==4.20.0
