+++ src/tests/test_login.py
@@ -1,0 +1,30 @@
+# Generated by Selenium IDE
+import pytest
+import time
+import json
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
+
+class TestLogin():
+  def setup_method(self, method):
+    self.driver = webdriver.Chrome()
+    self.vars = {}
+  
+  def teardown_method(self, method):
+    self.driver.quit()
+  
+  def test_login(self):
+    self.driver.get("http://127.0.0.1:5000/auth/login")
+    self.driver.set_window_size(1245, 1040)
+    self.driver.find_element(By.LINK_TEXT, "Log In").click()
+    self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+    self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+    self.driver.find_element(By.ID, "login-password").send_keys("password123")
+    self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+    self.driver.find_element(By.CSS_SELECTOR, ".alert-success").click()
+
